requirements.txt
chardet==5.2.0
python-dotenv==1.1.1
python-slugify==8.0.4
text-unidecode==1.3
